requirements.txt
torch
torchvision
timm
einops
numpy
pandas
matplotlib
seaborn
scikit-learn
tqdm
Pillow

streamlit
plotly
